feat: move burn form to portfolio asset detail dropdown
