+++ kit/dapp/src/app/[locale]/(private)/assets/[assettype]/[address]/(details)/_components/related/deposits.tsx
@@ -1,3 +1,4 @@
+import { BurnForm } from "@/app/[locale]/(private)/portfolio/my-assets/[assettype]/[address]/_components/burn-form/form";
 import { RelatedGrid } from "@/components/blocks/related-grid/related-grid";
 import { RelatedGridItem } from "@/components/blocks/related-grid/related-grid-item";
 import { getAssetBalanceDetail } from "@/lib/queries/asset-balance/asset-balance-detail";
@@ -6,7 +7,6 @@ import type { getDepositDetail } from "@/lib/queries/deposit/deposit-detail";
 import { isBefore } from "date-fns";
 import { getTranslations } from "next-intl/server";
 import type { Address } from "viem";
-import { BurnForm } from "../../../_components/manage-dropdown/burn-form/form";
 import { UpdateCollateralForm } from "../../../_components/manage-dropdown/update-collateral-form/form";
 import { MintForm } from "../../../_components/mint-form/form";
