+++ kit/dapp/src/app/[locale]/(private)/portfolio/my-assets/[assettype]/[address]/_components/burn-form/form.tsx
@@ -21,6 +21,7 @@ interface BurnFormProps {
   disabled?: boolean;
   asButton?: boolean;
   open?: boolean;
+  balance?: number;
   onOpenChange?: (open: boolean) => void;
 }
 
@@ -33,6 +34,7 @@ export function BurnForm({
   disabled = false,
   asButton = false,
   open,
+  balance,
   onOpenChange,
 }: BurnFormProps) {
   const t = useTranslations("private.assets.details.forms.form");
