+++ kit/dapp/src/app/[locale]/(private)/portfolio/my-assets/[assettype]/[address]/_components/burn-form/steps/summary.tsx
